feat(overview): mount as first dashboard tab
